Updated Grunt
I tried to speed up the grunt compilation by using concurrency. Now it
does sass and browserify at the same time. The problem is that normally
when we want a fast compilation we are only changing sass, so
concurrency does not matter. And it takes about 500ms to start a
worker.

Conclusion: concurrency slowed it down by a second or two, but I will
not remove it.

Also I moved all the grunt configs into files in the folder grunt/