I added to the white stuff
- I deleted the disclaimer on the sidebar,
- moved the sort to the side bar
- made the ask question in the side bar only if the user is not logged
in. Once the user is logged in there is an ask question button in the
action bar.
- I updated the whiteButton class in main.scss, added global colors.
- I think that constants should be all CAPS.
- I also made a toggle for sort, not done.
- put non intrusive grey for categories.
- Rewrote the author credit to “Asked x days ago by fin”
- Also changed the respond text, to a red button that gets the job done.
I also move the “luck” away from the top of the question list to the
side bar.

What was the motto about cards? I made the button in the side bar card
buttons.