Updated the question focus page.
I spend wayyyyy to long on trying to make it dynamically add a response
to the page when it is created.

I changed a lot of the layout and removed shadows.

I added a hidden delete button to the left and right of the responses.

Also the edit button is more subtle.